third_party/blink/web_tests/platform/generic/fast/dom/document-contentType-data-uri-expected.txt

PASS "application/xml" is "application/xml"
PASS "image/svg+xml" is "image/svg+xml"
PASS "text/html" is "text/html"
PASS "text/xml" is "text/xml"
PASS successfullyParsed is true

TEST COMPLETE


   
